feat: initial commit of delta zsh theme (edited from refined theme)
All credits to the original creators or the refined/pure theme and their origin themes. This theme at this stage is just my edits to make it simpler and easier to play with the colours so they match my VS Code themes (no shame in some good ol' colour coordination).